feat: add dialogue in ticksystem; fix theexpert in time;

+++ J.A.M/Assets/Editor/StorylineSystem/Data/Save/SSDialogueNodeSaveData.cs
@@ -10,7 +10,7 @@ namespace SS.Data.Save
     {
         [field: SerializeField] public string Text { get; set; }
         [field: SerializeField] public SSSpeakerType SpeakerType { get; set; }
-        [field: SerializeField] public uint Duration { get; set; }
+        [field: SerializeField] public float Duration { get; set; }
         [field: SerializeField] public bool IsDialogueTask { get; set; }
         [field: SerializeField] public int PercentageTask { get; set; }
         [field: SerializeField] public TraitsData.Job Job { get; set; }
